+++ CHANGES.md
@@ -1,6 +1,7 @@
 
 * Relicense from BSD3 to ISC.
 * Add `Fmt.Dump.signal` formatter for `Sys` signal numbers.
+* Add `Fmt[.Dump].result`, formatter for `result` values.
 * Add `Fmt.{words,paragraphs}` formatters on US-ASCII strings.
 * Add `Fmt.exn[_backtrace]`. Thanks to Edwin Török for suggesting.
 * Add `Fmt.quote`.